Remove the right-hand side correction for Dirichlet boundary conditions on offset grids
This correction works for one-dimensional problems, but in two dimensions and
more the right-hand side correction introduces an error since the right-hand
side then also contains second derivatives over other axes. Note that the
correction is not necessary to obtain second-order accuracy, but only works
to recover the exact solution (up to numerical precision) in the case that the
solution is given by a quadratic function (i.e. when the right-hand side is
constant). So instead of doing those corrections, I added a function that tests
whether a problem can be solved exactly for quadratic solutions. The tests
check whether this function is correct, i.e. test that the solution is almost
exact in the cases where it should be almost exact, and has significant error
(for coarse discretisations) otherwise.